# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Printly document printing platform.",
        "title": "Printly API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Kimba SABI N'GOYE",
            "email": "kimbasabingoye@printly.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all orders across the platform. Requires admin role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get all orders (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Order"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch all orders",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a single order by its ID. Requires admin role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get an order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Order"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch order",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an order. Requires admin role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Delete an order (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete order",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/centers/{id}/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all orders for a specific print center. Requires manager or admin role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Print Centers"
                ],
                "summary": "Get orders for a print center",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Print Center ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Order"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch orders",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new order with one or more documents and returns the created order. Requires authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Print Centers"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Print Center ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Order creation request",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Order"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Print center not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create order",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/status/{code}": {
            "get": {
                "description": "Retrieves the status of an order using its public pickup code.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order status by pickup code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pickup Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Order"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch order status",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the status of an order. Requires manager or admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Update an order's status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateOrderStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update status",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CreateOrderRequest": {
            "type": "object",
            "required": [
                "documents",
                "print_mode"
            ],
            "properties": {
                "documents": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.DocumentRequest"
                    }
                },
                "print_mode": {
                    "enum": [
                        "PRE_PRINT"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.PrintMode"
                        }
                    ]
                }
            }
        },
...
        "dto.DocumentRequest": {
            "type": "object",
            "required": [
                "file_name",
                "mime_type",
                "print_options",
                "size"
            ],
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "mime_type": {
                    "type": "string"
                },
                "print_options": {
                    "$ref": "#/definitions/entity.PrintOptions"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "A description of the error"
                }
            }
        },
        "dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
        "dto.UpdateOrderStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "$ref": "#/definitions/entity.OrderStatus"
                }
            }
        },
...
        "entity.ColorMode": {
            "type": "string",
            "enum": [
                "COLOR",
                "BLACK_AND_WHITE"
            ],
            "x-enum-varnames": [
                "Color",
                "BlackAndWhite"
            ]
        },
        "entity.Document": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mime_type": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                },
                "print_options": {
                    "$ref": "#/definitions/entity.PrintOptions"
                },
                "printed_at": {
                    "description": "When it was printed (nullable)",
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "storage_deleted_at": {
                    "description": "When it was deleted from GCS",
                    "type": "string"
                },
                "uploaded_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "entity.Order": {
            "type": "object",
            "properties": {
                "cancelled_at": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Document"
                    }
                },
                "id": {
                    "description": "gorm.Model is replaced to be explicit for swagger",
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                },
                "pickup_time": {
                    "type": "string"
                },
                "print_center_id": {
                    "type": "integer"
                },
                "print_mode": {
                    "$ref": "#/definitions/entity.PrintMode"
                },
                "status": {
                    "$ref": "#/definitions/entity.OrderStatus"
                },
                "user_uid": {
                    "type": "string"
                }
            }
        },
        "entity.OrderStatus": {
            "type": "string",
            "enum": [
                "CREATED",
                "AWAITING_DOCUMENT",
                "PENDING_PAYMENT",
                "PAID",
                "AWAITING_USER",
                "READY_TO_PRINT",
                "PRINTING",
                "PRINTED",
                "READY_FOR_PICKUP",
                "COMPLETED",
                "CANCELLED",
                "FAILED"
            ],
            "x-enum-varnames": [
                "StatusCreated",
                "StatusAwaitingDocument",
                "StatusPendingPayment",
                "StatusPaid",
                "StatusAwaitingUser",
                "StatusReadyToPrint",
                "StatusPrinting",
                "StatusPrinted",
                "StatusReadyForPickup",
                "StatusCompleted",
                "StatusCancelled",
                "StatusFailed"
            ]
        },
        "entity.PaperSize": {
            "type": "string",
            "enum": [
                "A4",
                "A3"
            ],
            "x-enum-varnames": [
                "A4",
                "A3"
            ]
        },
...
        "entity.PrintMode": {
            "type": "string",
            "enum": [
                "PRE_PRINT",
                "PRINT_UPON_ARRIVAL"
            ],
            "x-enum-varnames": [
                "PrePrint",
                "PrintUponArrival"
            ]
        },
        "entity.PrintOptions": {
            "type": "object",
            "properties": {
                "color": {
                    "description": "\"color\" | \"black_white\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.ColorMode"
                        }
                    ]
                },
                "copies": {
                    "type": "integer"
                },
                "pages": {
                    "description": "e.g., \"1-3,5\"",
                    "type": "string"
                },
                "paper_size": {
                    "description": "\"A4\", \"A3\", etc.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.PaperSize"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and a JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}